feat(sw): Service Worker throttling (buggy)
This is an attempt to have the service worker catch some request and throttle them (like a proxy).
Next step would have been to expose some kind of API (messaging or request) to change to throttling rate of the sw.

Though, fething urls protected by CORS result in an error (even for images), even if changing event.request.mode = "no-cors"
It seems that it needs to be done server-side 😢